refactor(extensions): move CodeParser to src/extensions/parser/, rename ext-babel → ext-parser-babel
- src/extensions/parser/{code-parser,index}.ts
- src/extensions/interfaces/index.ts shim re-exports from ../parser/
- extensions/ext-parser-babel (@veryfront/ext-parser-babel) maps veryfront/extensions/parser
- Root deno.json workspace + ./extensions/parser export
- Consumer stragglers (builtin-extensions.ts, code-parser-setup.ts, runtime name, tests) updated
