+++ deno.json
@@ -8,12 +8,12 @@
     "./extensions/ext-ai-google",
     "./extensions/ext-ai-openai",
     "./extensions/ext-auth-jwt",
-    "./extensions/ext-babel",
     "./extensions/ext-bundler-esbuild",
     "./extensions/ext-cache-redis",
     "./extensions/ext-mdx",
     "./extensions/ext-node-compat",
     "./extensions/ext-opentelemetry",
+    "./extensions/ext-parser-babel",
     "./extensions/ext-tailwind"
   ],
   "exclude": [
@@ -73,6 +73,7 @@
     "./extensions/auth": "./src/extensions/auth/index.ts",
     "./extensions/bundler": "./src/extensions/bundler/index.ts",
     "./extensions/cache": "./src/extensions/cache/index.ts",
+    "./extensions/parser": "./src/extensions/parser/index.ts",
     "./testing": "./src/testing/index.ts",
     "./testing/assert": "./src/testing/assert.ts",
     "./testing/bdd": "./src/testing/bdd.ts",
@@ -158,6 +159,7 @@
     "veryfront/extensions/bundler": "./src/extensions/bundler/index.ts",
     "veryfront/extensions/contracts": "./src/extensions/contracts.ts",
     "veryfront/extensions/interfaces": "./src/extensions/interfaces/index.ts",
+    "veryfront/extensions/parser": "./src/extensions/parser/index.ts",
     "jose": "npm:jose@5.9.6",
     "#veryfront": "./src/index.ts",
     "#veryfront/agent": "./src/agent/index.ts",
