+++ extensions/ext-parser-babel/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "@veryfront/ext-babel",
+  "name": "@veryfront/ext-parser-babel",
   "version": "0.1.0",
   "exports": "./src/index.ts",
   "veryfront": {
@@ -16,7 +16,7 @@
     "@std/assert": "jsr:@std/assert@1",
     "@std/testing/bdd": "jsr:@std/testing@1/bdd",
     "veryfront/extensions": "../../src/extensions/index.ts",
-    "veryfront/extensions/interfaces": "../../src/extensions/interfaces/index.ts"
+    "veryfront/extensions/parser": "../../src/extensions/parser/index.ts"
   },
   "tasks": {
     "test": "deno test --no-check --allow-all src/"
