+++ extensions/ext-parser-babel/src/inject-node-positions.ts
