+++ src/extensions/parser/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Parser category barrel — CodeParser (AST traversal) contract.
+ *
+ * @module extensions/parser
+ */
+
+export type {
+  ASTNode,
+  CodeParser,
+  GenerateOptions,
+  GenerateResult,
+  InjectJsxNodePositionsOptions,
+  NodePath,
+  ParseOptions,
+  TraverseVisitor,
+} from "./code-parser.ts";
