+++ src/transforms/plugins/babel-node-positions.ts
@@ -1,10 +1,10 @@
 /**
  * Shim over the `CodeParser` contract's `injectJsxNodePositions()` method.
  *
- * The actual Babel-based implementation lives in `@veryfront/ext-babel`.
+ * The actual Babel-based implementation lives in `@veryfront/ext-parser-babel`.
  * Core modules continue to import from this path; the shim resolves the
  * contract at call time. Position injection powers Studio Navigator only —
- * if `ext-babel` is not installed we return the source unchanged rather
+ * if `ext-parser-babel` is not installed we return the source unchanged rather
  * than fail the SSR path.
  */
